requirements.txt
nextcord~=2.6.0
python-dotenv~=1.0.0
discord~=2.3.2
flask~=3.0.3